+++ benches/note_decryption.rs
@@ -5,7 +5,7 @@ use orchard::{
     circuit::ProvingKey,
     keys::{FullViewingKey, PreparedIncomingViewingKey, Scope, SpendingKey},
     note::AssetBase,
-    note_encryption_v3::{CompactAction, OrchardDomainV3},
+    note_encryption_orchardzsa::{CompactAction, OrchardZSADomain},
     value::NoteValue,
     Anchor, Bundle,
 };
@@ -79,7 +79,7 @@ fn bench_note_decryption(c: &mut Criterion) {
     };
     let action = bundle.actions().first();
 
-    let domain = OrchardDomainV3::for_action(action);
+    let domain = OrchardZSADomain::for_action(action);
 
     let compact = {
         let mut group = c.benchmark_group("note-decryption");
@@ -120,12 +120,12 @@ fn bench_note_decryption(c: &mut Criterion) {
         let ivks = 2;
         let valid_ivks = vec![valid_ivk; ivks];
         let actions: Vec<_> = (0..100)
-            .map(|_| (OrchardDomainV3::for_action(action), action.clone()))
+            .map(|_| (OrchardZSADomain::for_action(action), action.clone()))
             .collect();
         let compact: Vec<_> = (0..100)
             .map(|_| {
                 (
-                    OrchardDomainV3::for_action(action),
+                    OrchardZSADomain::for_action(action),
                     CompactAction::from(action),
                 )
             })
