+++ src/lib.rs
@@ -27,7 +27,9 @@ pub mod keys;
 pub mod note;
 pub mod supply_info;
 // pub mod note_encryption; // disabled until backward compatability is implemented.
-pub mod note_encryption_v3;
+pub mod note_encryption_orchardzsa;
+
+mod note_encryption_orchard;
 pub mod primitives;
 mod spec;
 pub mod tree;
